Adding social media window to the sophia app.
twitterConnect.js needs to be run on the server in node with vaild twitter application id and key
in order to act as a relay between the web app and twitter.

To add to twitterConnect.js - some filtering of acceptable hash tag searches to avoid abuse of given
that users do not use any authentication to log into sophia.

Changes to be committed:
	modified:   app/index.html
	new file:   app/socialImage.jpg
	modified:   config.yml
	modified:   src/img/icons.svg
	modified:   src/js/resources/en.js
	new file:   src/js/windows/social.js
	modified:   src/sass/app.scss
	new file:   src/sass/windows/socialwindow.scss
	new file:   twitter/twitterConnect.js